requirements.txt
pip==23.3.2
python-generate-mac==1.3.1
optparse-pretty==0.1.1
regex-toolkit==0.0.5
subprocess.run==0.0.8